Separate test concerns
