+++ ckanext/string_to_location/tests/test_plugin.py
@@ -22,9 +22,6 @@ import ckan.tests.helpers as helpers
 from ckan.lib.helpers import url_for
 from ckan.common import config
 
-from ckanext.string_to_location.location_mapper_log_reader import LocationMapperLogReader
-from ckanext.string_to_location.location_mapper_log_writer import LocationMapperLogWriter
-
 
 class TestString_To_LocationPlugin(object):
 
@@ -43,66 +40,6 @@ class TestString_To_LocationPlugin(object):
         model.repo.rebuild_db()
         config.update(cls.original_config)
         ckan.plugins.unload('string_to_location')
-
-    def test_write_info_log_message_for_a_resource(self):
-        user, resource, context = self._create_context()
-
-        log_writer = LocationMapperLogWriter(resource['id'])
-        log_writer.info("Testing", context=context)
-
-        task = ckan.plugins.toolkit.get_action('task_status_show')(context, {
-            'entity_id': resource['id'],
-            'task_type': 'location_mapper',
-            'key': 'location_mapper'
-        })
-
-        value = json.loads(task['value'])
-
-        assert_true(value['logs'][0]['message'], "Testing")
-
-    def test_write_warn_log_message_for_a_resource(self):
-        user, resource, context = self._create_context()
-
-        log_writer = LocationMapperLogWriter(resource['id'])
-        log_writer.warn("Warning", context=context)
-
-        task = ckan.plugins.toolkit.get_action('task_status_show')(context, {
-            'entity_id': resource['id'],
-            'task_type': 'location_mapper',
-            'key': 'location_mapper'
-        })
-
-        value = json.loads(task['value'])
-
-        assert_true(value['logs'][0]['message'], "Warning")
-
-    def test_write_error_log_message_for_a_resource(self):
-        user, resource, context = self._create_context()
-
-        log_writer = LocationMapperLogWriter(resource['id'])
-        log_writer.error("Oops", context=context)
-
-        task = ckan.plugins.toolkit.get_action('task_status_show')(context, {
-            'entity_id': resource['id'],
-            'task_type': 'location_mapper',
-            'key': 'location_mapper'
-        })
-
-        value = json.loads(task['value'])
-
-        assert_true(value['logs'][0]['message'], "Oops")
-
-    def test_read_log_message_for_a_resource(self):
-        user, resource, context = self._create_context()
-
-        log_writer = LocationMapperLogWriter(resource['id'])
-        log_writer.info("Testing", context=context)
-
-        log_reader = LocationMapperLogReader(resource['id'])
-        mapper_status = log_reader.get_status()
-
-        assert_true(mapper_status['task_info'][
-                    'logs'][0]['message'], "Testing")
 
     def test_map_location_with_correctly_formatted_file(self):
         app = helpers.FunctionalTestBase._get_test_app()
@@ -188,15 +125,7 @@ class TestString_To_LocationPlugin(object):
         response = helpers.webtest_maybe_follow(app.get("/dataset/" + package['id'] + "/resource/" + resource['id']+ "/map_location",
             extra_environ={'REMOTE_USER': str(user['name'])})) 
 
-        response.mustcontain("The resource does not specify location columns")   
-    
-    def test_map_location_with_incorrectly_formatted_file(self):
-        # FIXME add test when we have handling for incorrectly formatted files
-        pass
-        
-    def test_map_location_with_incorrectly_formatted_file_no_uploads_to_dataset(self):
-        # FIXME add test when we have handling for incorrectly formatted files
-        pass
+        response.mustcontain("The resource does not specify location columns")
 
     def _create_context(self):
         user = factories.Sysadmin()
